Time: 0 ms (100%), Space: 8.6 MB (37.75%) - LeetHub

+++ 0029-divide-two-integers/0029-divide-two-integers.cpp
@@ -1,6 +1,6 @@
 class Solution {
 public:
-    int divide(int dividend, int divisor) {
+    int divide(long long dividend, long long divisor) {
         if(dividend == INT_MIN && divisor ==-1) return INT_MAX;
         return dividend/divisor;
     }
